Создает абстрактный класс, рефакторит код под абстракции, переносит логику в контроллер
