+++ src/components/no-films.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component.js";
 
 const createNoDataTemplate = () => {
   return (
@@ -10,24 +10,8 @@ const createNoDataTemplate = () => {
   );
 };
 
-export default class NoData {
-  constructor() {
-    this._element = null;
-  }
-
+export default class NoData extends AbstractComponent {
   getTemplate() {
     return createNoDataTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
