+++ src/components/show-more.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component.js";
 
 const createShowMoreTemplate = () => {
   return (
@@ -6,24 +6,12 @@ const createShowMoreTemplate = () => {
   );
 };
 
-export default class ShowMore {
-  constructor() {
-    this._element = null;
-  }
-
+export default class ShowMore extends AbstractComponent {
   getTemplate() {
     return createShowMoreTemplate();
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
+  setClickHandler(handler) {
+    this.getElement().addEventListener(`click`, handler);
   }
 }
