+++ src/components/statistics.js
@@ -1,30 +1,17 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component.js";
 
 export const createStatisticsTemplate = (allMovies) => {
   return (
     `<p>${allMovies} movies inside</p>`
   );
 };
-export default class Statistics {
+export default class Statistics extends AbstractComponent {
   constructor(allMovies) {
+    super();
     this._allMovies = allMovies;
-
-    this._element = null;
   }
 
   getTemplate() {
     return createStatisticsTemplate(this._allMovies);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
