+++ src/utils/common.js
@@ -1,7 +1,3 @@
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
 
 const transformTimeFormat = (filmTime) => {
   const filmHours = Math.floor(filmTime / 60);
@@ -48,35 +44,17 @@ const getArrayFromText = (text) => {
 
 const getFavoriteGenre = (arr) => {
   let genres = arr
-    .filter((movie) => movie.isInFavorites)
-    .map(({genre}) => genre)
-    .flat()
-    .reduce((obj, genre) => {
-      obj[genre] = obj[genre] ? ++obj[genre] : 1;
-      return obj;
-    }, {});
+      .filter((movie) => movie.isInFavorites)
+      .map(({genre}) => genre)
+      .flat()
+      .reduce((obj, genre) => {
+        obj[genre] = obj[genre] ? ++obj[genre] : 1;
+        return obj;
+      }, {});
   const arrGenres = Object.entries(genres).sort((a, b) => b[1] - a[1]);
   const maxGenre = arrGenres[0][1];
   return arrGenres.filter((genre) => genre[1] === maxGenre).map((genre) => genre[0]).join(`, `);
 };
-
-// const getFavoriteGenre = (arr) => {
-//   let genreItem = arr[0];
-//   let maxRepeatNumber = 1;
-//   for (let i = 0; i < arr.length; i++) {
-//     let repeatNumber = 0;
-//     for (let j = 0; j - arr.length - 1; j++) {
-//       if (arr[i] === arr[j]) {
-//         repeatNumber += 1;
-//       }
-//     }
-//     if (repeatNumber > maxRepeatNumber) {
-//       maxRepeatNumber = repeatNumber;
-//       genreItem = arr[i];
-//     }
-//   }
-//   return genreItem;
-// };
 
 const getUserRating = (arr) => {
   const watchedMoviesQuantity = arr.length;
@@ -91,24 +69,5 @@ const getUserRating = (arr) => {
   return userRating;
 };
 
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-const render = (container, element, place = RenderPosition.BEFOREEND) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
 export {transformTimeFormat, getFavoriteGenre, getUserRating, getRandomDate, getRandomArrayItem,
-  getRandomNumber, getRandomFractionalNumber, makeRandomArr, getArrayFromText, createElement, render,
-  RenderPosition};
+  getRandomNumber, getRandomFractionalNumber, makeRandomArr, getArrayFromText};
